Added pygame.mixer.music support
Now the user can choose whether to get the audio of the video as a pygame.mixer.Sound or just use pygame.mixer.music. This was mainly for audio syncing purposes, since with mixer.music you can know exactly where the audio is at (so if any desync occurs, it can be resynced).